split http server into 2 classes: request router and actual http server to prepare new routing with namespaces #10
